Add Huggingface benchmark
