--- crates/bpe/script/copy-benchmark-results
@@ -1,11 +1,0 @@
-#!/usr/bin/env bash
-
-set -eu
-
-result_dir="benches/result"
-
-mkdir -p "$result_dir"
-
-for i in {counting,encoding,appending,worstcase}-o200k; do
-    rsvg-convert --format svg --output "$result_dir/$i.svg" --background-color white "target/criterion/reports/$i/lines.svg"
-done
